refactor(about_app): move RPC functions to ops module for better organization
- Moved list_capabilities, lookup_capability, and search_capabilities functions from about_app/mod.rs to a new ops.rs module.
- Updated exports in mod.rs to reflect the new structure, improving code organization and maintainability.
